Extract Container extend tests
